CMake improved, Editor warnings as errors

+++ Editor/src/UI/Panels/GameViewPanel.cpp
@@ -4,6 +4,10 @@
 
 #include "Twisted/Rendering/OpenGL/FrameBuffer_OpenGL.h"
 #include "Twisted/Gameing/GameService.h"
+#include "Twisted/Rendering/FrameBuffer.h"
+#include "Twisted/TObject.h"
+#include "UI/EditorPanel.h"
+#include <imgui.h>
 
 namespace Twisted::Editor
 {
@@ -28,7 +32,7 @@ namespace Twisted::Editor
 		if (fb && fb->Tex)
 			ImGui::Image(
 				(void*)(intptr_t)fb->Tex->TexID,
-				ImVec2(Size.x, Size.y),
+				ImVec2((float)Size.x, (float)Size.y),
 				ImVec2(0, 1),  // top-left UV
 				ImVec2(1, 0)   // bottom-right UV (flipped vertically)
 			);
